+++ plugins/splunk-on-call/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-splunk-on-call
 
+## 0.3.30-next.1
+
+### Patch Changes
+
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/core-components@0.9.5-next.1
+  - @backstage/core-plugin-api@1.0.3-next.0
+  - @backstage/catalog-model@1.0.3-next.0
+  - @backstage/plugin-catalog-react@1.1.1-next.1
+
 ## 0.3.30-next.0
 
 ### Patch Changes